Move CA1060 to it's new location in Microsoft.ApiGuidelines analyzer
