+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/UnitTests/Microsoft.ApiDesignGuidelines.Analyzers.UnitTests.csproj
@@ -218,6 +218,9 @@
     <Compile Include="ConstructorMakeNoninheritableBaseClassInheritableTests.cs" />
     <Compile Include="ConstructorMakeNoninheritableBaseClassInheritableTests.Fixer.cs" />
   </ItemGroup>
+  <ItemGroup>
+    <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
+  </ItemGroup>
   <ImportGroup Label="Targets">
     <Import Project="..\..\..\build\Targets\Analyzers.Imports.targets" />
   </ImportGroup>
