+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/VisualBasic/Microsoft.ApiDesignGuidelines.VisualBasic.Analyzers.vbproj
@@ -129,7 +129,6 @@
     <Compile Include="BasicStaticHolderTypesShouldBeSealed.Fixer.vb" />
     <Compile Include="BasicStaticHolderTypesShouldNotHaveConstructors.vb" />
     <Compile Include="BasicStaticHolderTypesShouldNotHaveConstructors.Fixer.vb" />
-    <Compile Include="BasicMovePInvokesToNativeMethodsClass.vb" />
     <Compile Include="BasicMovePInvokesToNativeMethodsClass.Fixer.vb" />
     <Compile Include="BasicIdentifiersShouldDifferByMoreThanCase.vb" />
     <Compile Include="BasicIdentifiersShouldDifferByMoreThanCase.Fixer.vb" />
