+++ PortFxCop/src/System.Runtime.InteropServices.Analyzers/UnitTests/System.Runtime.InteropServices.Analyzers.UnitTests.csproj
@@ -158,6 +158,9 @@
     <Compile Include="UseManagedEquivalentsOfWin32ApiTests.Fixer.cs" />
     <Compile Include="AlwaysConsumeTheValueReturnedByMethodsMarkedWithPreserveSigAttributeTests.cs" />
   </ItemGroup>
+  <ItemGroup>
+    <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
+  </ItemGroup>
   <ImportGroup Label="Targets">
     <Import Project="..\..\..\build\Targets\Analyzers.Imports.targets" />
   </ImportGroup>
